Added hooks for gathering mods.
